// src/app/models/user.model.ts
/*
 * Copyright (C) 2007-2019 Crafter Software Corporation. All Rights Reserved.
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 *
 * You should have received a copy of the GNU General Public License
 * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */

export interface User {
  id?: number;
  email: string;
  username: string;
  firstName: string;
  lastName: string;
  enabled: boolean;
  password?: string;
  // groups: Array<number>;
  // expired: boolean;
  // locked: boolean;
  // credentialsExpired: boolean;
  authenticationType: string;
  externallyManaged: boolean;
  avatarUrl: string;
}